+++ app/controllers/resources/access_policies_controller.rb
@@ -4,26 +4,37 @@ module Resources
     before_action :find_and_authorize
 
     def create
-      @access_policy = AccessPolicy.new(permitted_attributes(AccessPolicy))
+      @access_policy = @resource.access_policies.
+                       build(permitted_attributes(AccessPolicy))
 
-      if @access_policy.save
-        redirect_to service_global_policy_path(@resource.service, @resource)
-      else
-        @model = ResourceAccessPoliciesDecorator.
-                 new(current_user, @resource, @access_policy)
-        @service = @resource.service
-        render('services/global_policies/show', status: :bad_request)
-      end
+      @access_policy.save ? redirect_to_list : render_errors
     end
 
     def destroy
       access_policy = AccessPolicy.find(params[:id])
       access_policy.destroy!
 
-      redirect_to service_global_policy_path(@resource.service, @resource)
+      redirect_to_list
     end
 
     private
+
+    def redirect_to_list
+      if @resource.global?
+        redirect_to service_global_policy_path(@resource.service, @resource)
+      else
+        redirect_to service_local_policy_path(@resource.service, @resource)
+      end
+    end
+
+    def render_errors
+      @model = ResourceAccessPoliciesDecorator.
+               new(current_user, @resource, @access_policy)
+      @service = @resource.service
+
+      render("services/#{@resource.resource_type}_policies/show",
+             status: :bad_request)
+    end
 
     def find_and_authorize
       @resource = Resource.find(params[:resource_id])