+++ spec/factories/resources.rb
@@ -10,6 +10,11 @@ FactoryGirl.define do
       resource_type :global
     end
 
+    trait :local do
+      resource_type :local
+    end
+
     factory :global_resource, traits: [:global]
+    factory :local_resource, traits: [:local]
   end
 end