+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Pundit authorized error messages for groups and services (@mkasztelnik)
 - Notifications are using the JS toastr library for fancier popups (@dharezlak)
 - The file store component uses portal's notification system to report errors (@dharezlak)
+- Service owner can manage local policies through UI (@mkasztelnik)
 
 ### Changed
 - Redirect status set to 302 instead of 404 (when record not found), 401 (when